split type into individual files under ./types
this is to avoid runtime requires, that browserify would not be able to handle.
