+++ lib/schema.js
@@ -1,5 +1,5 @@
 var es = require('event-stream');
-var Document = require('./type').Document;
+var Document = require('./types').Document;
 
 /**
  * The top level schema document, like a Document type
