+++ lib/type-collection.js
@@ -3,6 +3,8 @@ var lodashMixin = require('ampersand-collection-lodash-mixin');
 var parentMixin = require('./collection-parent-mixin');
 var assert = require('assert');
 var _ = require('lodash');
+var Type = require('./types/type');
+var types = require('./types');
 var debug = require('debug')('mongodb-schema:type-collection');
 
 /**
@@ -37,13 +39,13 @@ module.exports = Collection.extend(lodashMixin, parentMixin, {
   },
   model: function(attrs, options) {
     // require ./type at runtime to avoid circular imports, does this work with browserify??
-    var Klass = require('./type')[attrs.name];
+    var Klass = types[attrs.name];
     assert(Klass, 'No value type for ' + attrs.name);
     return new Klass(attrs, options);
   },
   isModel: function(model) {
     // require ./type at runtime to avoid circular imports, does this work with browserify??
-    return (model instanceof require('./type').Type);
+    return (model instanceof Type);
   },
   /**
    * adds a new value to the correct type, and creates the type first
