+++ lib/types/index.js
@@ -1,0 +1,8 @@
+var _ = require('lodash');
+
+_.extend(module.exports,
+  require('./constant'),
+  require('./primitive'),
+  require('./array'),
+  require('./document')
+);
